make the project maven to ease usage of dependencies
