--- .classpath
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.7"/>
-	<classpathentry kind="lib" path="libs/java-client-2.2.0.jar"/>
-	<classpathentry kind="lib" path="libs/jxl.jar"/>
-	<classpathentry kind="lib" path="libs/saxon-8.7.jar"/>
-	<classpathentry kind="lib" path="libs/SaxonLiaison.jar"/>
-	<classpathentry kind="lib" path="libs/selenium-java-2.52.0.jar"/>
-	<classpathentry kind="lib" path="libs/selenium-server-standalone-2.52.0.jar"/>
-	<classpathentry kind="lib" path="libs/snakeyaml-1.13.jar"/>
-	<classpathentry kind="lib" path="libs/testng-6.9.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
