+++ src/main/java/com/github/priyanshus/automation/sdk/Constants.java
@@ -1,0 +1,15 @@
+package com.github.priyanshus.automation.sdk;
+
+/**
+ * Constants of project
+ */
+public class Constants {
+    public static final String SCREENSHOT_BASEPATH = System
+        .getProperty("user.dir") + "/screenshots";
+
+    public static final String SEARCH_BASEPATH = System.getProperty("user.dir")
+        + "/searchimages";
+
+    public static final String DIFF_BASEPATH = System.getProperty("user.dir")
+        + "/diffimages";
+}
